reader/requirements.txt
py7zr==0.22.0
pyzipper==0.3.6
rarfile==4.2
PyYAML==6.0.2